Clarify memory measurements
The precise meaning of the tim.exe memory measurement is not obvious and
this ambiguity makes it less useful. While doing ninja measurements I
couldn't tell if it referred to physical memory or commit (it's commit)
and I couldn't tell if it referred to the peak of a single process or of
the process tree (single process).

This updates the README and adds code comments to clarify this.